Change Mod+Tab to cycle through clients
This instead of focusing back on the last client.
Since I can use tags in awesome, it better for me to not have too many
client opened at one time on the same tag.
Since I use Mod+something for a lot of navigation, having to shift to
using Alt+Tab for cycling is not something I do intuitively. Thus the
change